# src/main/webapp/docs/swagger.yaml
openapi: 3.0.0
servers: 
  - url: 'https://api-electronicdevices-313314.nw.r.appspot.com/api/v1/'
  - url: 'http://localhost:8089/api/v1/'
info:
  description: This API can be used to consult multiple electronic devices which could be found in different shops.
  version: 2.0.0
  title: API-ElectronicDevices
  contact:
    email: joselugarciamarin2406@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /accessories:
    get:
      tags:
        - Accessories
      summary: Get accessories
      description: Get all the accessories
      operationId: getAccesories
      parameters:
        - name: order
          in: query
          description: Order results by name or price
          required: false
          schema:
            type: string
            enum:
              - name
              - '-name'
              - price
              - '-price'
        - name: isEmpty
          in: query
          description: Filter accessories with no devices
          required: false
          schema:
            type: boolean
        - name: name
          in: query
          description: Filter accessories by name
          required: false
          schema: 
            type: string
        - name: wireless
          in: query
          description: Filter accessories by wireless
          schema:
            type: string
            enum:
            - 'true'
            - 'false'
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accessory'
        '400':
          description: Bad request
    post:
      tags:
        - Accessories
      summary: Create a accessory
      description: Create a accessory
      operationId: createAccessory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessoryEmpty'
        description: Accessory to create
        required: true
      responses:
        '201':
          description: Accessory created
        '400':
          description: Bad request
    put:
      tags:
        - Accessories
      summary: Update an accessory
      description: Upadate an accessory
      operationId: updateAccessory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Accessory'
        description: Accessory to update
        required: true
      responses:
        '204':
          description: Accessory updated
        '400':
          description: Bad request
  '/accessories/{accessoryId}/{deviceId}':
    post:
      tags: 
        - Accessories
      summary: Add a device compatible with an accessory
      description: Add a device compatible with an accessory
      operationId: addDeviceToAccessories
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      parameters: 
        - name: accessoryId
          in: path
          description: Id of the accessory
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: deviceId
          in: path
          description: Id of the device
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '201':
          description: Device added to the compatible devices list with an accesory
        '400':
          description: Bad request
        '404':
          description: Accessory or device not found
    delete:
      tags: 
        - Accessories
      summary: Delete a device compatible with an accessory
      description: Delete a device compatible with an accessory
      operationId: deleteDeviceFromAccesory
      parameters: 
        - name: accessoryId
          in: path
          description: Id of the accessory
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: deviceId
          in: path
          description: Id of the device
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Device added to the compatible devices list with an accesory 
        '404':
          description: Accessory or device not found
  '/accessories/{id}':
    get:
      tags:
        - Accessories
      summary: Get an accessory by id
      description: Get an accessory  by id
      operationId: getAccessory
      parameters:
        - name: id
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses: 
        '200':
          description: Accessory found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accessory'
        '400':
          description: Accessory not found
    delete:
      tags:
        - Accessories
      summary: Delete an accessory
      description: Delete an accessory
      operationId: deleteAccessory
      parameters:
        - name: id
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Accessory deleted
        '404':
          description: Accessory not found
...
components:
  schemas:
...
    Accessory:
      type: array
      items: 
        type: object
        required:
          - name
        properties:
          accessoryId:
            type: integer
          name: 
            type: string
          wireless: 
            type: boolean
          price:
            type: number
          compatibleDevices:
            type: array
            items: 
              type: object
              required: 
                - name
              properties:
                id:
                  type: integer
                name:
                  type: string
                brand:
                  type: string
                price: 
                  type: number
                screenSize:
                  type: number
                weight:
                  type: number
                deviceType:
                  type: string
                storage:
                  type: integer
                processor:
                  type: string
                memory:
                  type: integer
                OS:
                  type: string
                setColors:
                  type: array
                  items:
                    type: string
                camera:
                  type: string
      example:
        accessoryId: 27
        name: Logitec MX Master 3
        wireless: true
        price: 89.99
        compatibleDevice:
          id: 6
          name: Rog Strix G15
          brand: Asus
          price: 1490.0
          screenSize: 15.6
          weight: 2.5
          deviceType: COMPUTER
          storage: 1000
          processor: Intel i7-10750H
          memory: 16
          setColors: [Black, Grey]
          camera: "-"
          os: Windows 10
    AccessoryEmpty:
      type: array
      items: 
        type: object
        required:
          - name
        properties:
          accessoryId:
            type: integer
          name: 
            type: string
          wireless: 
            type: boolean
          price:
            type: number
          compatibleDevices:
            type: array
            items:
              type: object
      example:
        accessoryId: 27
        name: Logitec MX Master 3
        wireless: true
        price: 89.99
        compatibleDevice:
          id: 6
          name: Rog Strix G15
          brand: Asus
          price: 1490.0
          screenSize: 15.6
          weight: 2.5
          deviceType: COMPUTER
          storage: 1000
          processor: Intel i7-10750H
          memory: 16
          setColors: [Black, Grey]
          camera: "-"
          os: Windows 10